refactor(querycursor,aggregate): move eachAsync into helper
Re: #4244
