+++ lib/aggregate.js
@@ -6,6 +6,7 @@ var util = require('util');
 var utils = require('./utils');
 var PromiseProvider = require('./promise_provider');
 var Query = require('./query');
+var eachAsync = require('./services/cursor/eachAsync');
 var read = Query.prototype.read;
 
 /**
@@ -675,48 +676,7 @@ Aggregate.prototype.exec = function(callback) {
 
 function decorateCursor(cursor) {
   cursor.eachAsync = function(fn, callback) {
-    var Promise = PromiseProvider.get();
-
-    var handleNextResult = function(doc, callback) {
-      var promise = fn(doc);
-      if (promise && typeof promise.then === 'function') {
-        promise.then(
-          function() { callback(null); },
-          function(error) { callback(error); });
-      } else {
-        callback(null);
-      }
-    };
-
-    var iterate = function(callback) {
-      return cursor.next(function(error, doc) {
-        if (error) {
-          return callback(error);
-        }
-        if (!doc) {
-          return callback(null);
-        }
-        handleNextResult(doc, function(error) {
-          if (error) {
-            return callback(error);
-          }
-          setTimeout(function() {
-            iterate(callback);
-          }, 0);
-        });
-      });
-    };
-
-    return new Promise.ES6(function(resolve, reject) {
-      iterate(function(error) {
-        if (error) {
-          callback && callback(error);
-          return reject(error);
-        }
-        callback && callback(null);
-        return resolve();
-      });
-    });
+    return eachAsync(function(cb) { return cursor.next(cb); }, fn, callback);
   };
 }
