+++ lib/querycursor.js
@@ -4,6 +4,7 @@
 
 var PromiseProvider = require('./promise_provider');
 var Readable = require('stream').Readable;
+var eachAsync = require('./services/cursor/eachAsync');
 var helpers = require('./queryhelpers');
 var util = require('util');
 
@@ -190,50 +191,9 @@ QueryCursor.prototype.next = function(callback) {
  */
 
 QueryCursor.prototype.eachAsync = function(fn, callback) {
-  var Promise = PromiseProvider.get();
   var _this = this;
 
-  var handleNextResult = function(doc, callback) {
-    var promise = fn(doc);
-    if (promise && typeof promise.then === 'function') {
-      promise.then(
-        function() { callback(null); },
-        function(error) { callback(error); });
-    } else {
-      callback(null);
-    }
-  };
-
-  var iterate = function(callback) {
-    return _next(_this, function(error, doc) {
-      if (error) {
-        return callback(error);
-      }
-      if (!doc) {
-        return callback(null);
-      }
-      handleNextResult(doc, function(error) {
-        if (error) {
-          return callback(error);
-        }
-        // Make sure to clear the stack re: gh-4697
-        setTimeout(function() {
-          iterate(callback);
-        }, 0);
-      });
-    });
-  };
-
-  return new Promise.ES6(function(resolve, reject) {
-    iterate(function(error) {
-      if (error) {
-        callback && callback(error);
-        return reject(error);
-      }
-      callback && callback(null);
-      return resolve();
-    });
-  });
+  return eachAsync(function(cb) { return _next(_this, cb); }, fn, callback);
 };
 
 /*!
